Refactor UI elements to UI_menus. Similarly refactor tests. Rename files appropriately. Group UI scripts (eg from chart_generator) together within UI_menus.
Move display_save_image_save_as from UI_functions to
UI_menus/chart_generator/create_chart_UI.

Add [Issue #124](https://github.com/toonarmycaptain/dionysus/issues/124)
to move created chart display and saving to from generate_image to
create_chart as these functions don't relate to image generation and
shouldn't be the responsibility of generate_image.py
