+++ dionysus_app/UI_menus/UI_functions.py
@@ -4,10 +4,7 @@ UI functions: user interface functions used throughout the application.
 
 import tkinter as tk
 
-from pathlib import Path
 from tkinter import filedialog
-
-from PIL import ImageTk, Image
 
 
 def clear_screen(num_lines=50):
@@ -192,56 +189,6 @@ def select_folder_dialogue(title_str=None, start_dir='..'):
     if dir_path_str == '':
         return None
     return dir_path_str
-
-
-def display_image_save_as(chart_image_path: str):
-    """
-    Displays the given image path in a window titled with the filename.
-    Supplies a button "Save as" below, which quits the window. Save as
-    dialogue is called separately (ideally directly subsequently, from
-    a UI perspective.
-
-    :param chart_image_path: str or Path object.
-    :return: None
-    """
-    root = tk.Tk()
-    root.title(Path(chart_image_path).name)  # Adding Path() allows string or Path object.
-    root.geometry("960x580")  # Half dimensions, about a quarter of full image.
-
-    image_display = ImageDisplay(image_path=chart_image_path, master=root)
-    image_display.mainloop()
-
-    root.destroy()
-
-
-class ImageDisplay(tk.Frame):
-    def __init__(self, image_path=None, master=None):
-        super().__init__(master)
-        self.master = master
-
-        if image_path:
-            self.image_path = image_path
-            self.pack()
-            self.create_image_and_button()
-        else:
-            pass  # don't instantiate?
-
-    def create_image_and_button(self):
-        self.create_image_widget()
-        self.create_save_as_button_widget()
-
-    def create_image_widget(self):
-        self.full_chart_image = Image.open(self.image_path)
-        self.display_sized_image = self.full_chart_image.resize((960, 540), Image.ANTIALIAS)
-
-        self.display_image = ImageTk.PhotoImage(self.display_sized_image, master=self.master)
-        self.image_panel = tk.Label(self, image=self.display_image)
-        self.image_panel.pack(side="top", fill="both", expand="yes")
-
-    def create_save_as_button_widget(self):
-        self.quit = tk.Button(self, text="Save as", font=('Arial', 24),
-                              command=self.quit)
-        self.quit.pack(side="bottom")
 
 
 if __name__ == '__main__':
