+++ dionysus_app/UI_menus/main_menu.py
@@ -6,7 +6,7 @@ import sys
 
 from dionysus_app.class_functions import create_classlist
 from dionysus_app.chart_generator.create_chart import new_chart
-from dionysus_app.UI_menus.edit_class_data import edit_class_data
+from dionysus_app.UI_menus.edit_class_data_UI import edit_class_data
 from dionysus_app.UI_menus.settings_menu import run_settings_menu
