+++ dionysus_app/chart_generator/create_chart.py
@@ -13,9 +13,9 @@ from dionysus_app.class_functions import select_classlist
 from dionysus_app.data_folder import CHART_DATA_FILE_TYPE, DataFolder
 from dionysus_app.chart_generator.generate_image import generate_chart_image
 from dionysus_app.chart_generator.process_chart_data import DEFAULT_CHART_PARAMS
-from dionysus_app.chart_generator.take_chart_data import take_chart_name, take_score_data
 from dionysus_app.file_functions import convert_to_json
 from dionysus_app.UI_menus.UI_functions import clean_for_filename
+from dionysus_app.UI_menus.chart_generator.take_chart_data_UI import take_chart_name, take_score_data
 
 
 CLASSLIST_DATA_PATH = DataFolder.generate_rel_path(DataFolder.CLASS_DATA.value)
