+++ test_suite/UI_menus_tests/test_edit_class_data_UI.py
@@ -1,4 +1,4 @@
-from dionysus_app.UI_menus.edit_class_data import edit_class_data
+from dionysus_app.UI_menus.edit_class_data_UI import edit_class_data
 
 
 def test_edit_class_data():
