+++ test_suite/UI_menus_tests/test_take_chart_data.py
